requirements.txt
django>=5.0,<6.0
dj-database-url
gunicorn
tzdata
pillow>=10.0.0
django-environ>=0.11.0,<1.0
djangorestframework-simplejwt>=5.3.0
djangorestframework>=3.14.0
drf-yasg>=1.21.0
django-cors-headers>=4.3.0
requests>=2.31.0

